--- a/Cargo.toml
+++ b/Cargo.toml
@@ -19,3 +19,12 @@ winit = "0.30.3"
 wgpu = "23"
 resvg = { version = "0.44", default-features = false }
 pollster = "0.4.0"
+criterion = { version = "0.5", features = ["html_reports"] }
+
+[profile.profiling]
+inherits = "release"
+debug = true
+
+[[bench]]
+name = "prepare"
+harness = false