--- a/Cargo.toml
+++ b/Cargo.toml
@@ -21,10 +21,6 @@ resvg = { version = "0.44", default-features = false }
 pollster = "0.4.0"
 criterion = { version = "0.5", features = ["html_reports"] }
 
-[profile.profiling]
-inherits = "release"
-debug = true
-
 [[bench]]
 name = "prepare"
 harness = false