--- a/samples/README.md
+++ b/samples/README.md
@@ -1,0 +1,8 @@
+# Sample Attribution
+
+The sample sources with accompanying attributions and licenses are as follows:
+
+| Sample         | Title                       | Amendments | Source                                                                                                                                                               | License                                                                                            |
+| -------------- | --------------------------- | ---------- | -------------------------------------------------------------------------------------------------------------------------------------------------------------------- | -------------------------------------------------------------------------------------------------- |
+| `./arabic.txt` | Al-Kindi - First Philosophy | None       | [Wiki Source](<https://en.wikisource.org/wiki/Moby-Dick_(1851)_US_edition/Chapter_1>)                                                                                | [Creative Commons Attribution-ShareAlike License](https://creativecommons.org/licenses/by-sa/4.0/) |
+| `./latin.txt`  | Moby Dick - First Chapter   | None       | [Wiki Source](https://ar.wikisource.org/wiki/%D8%A7%D9%84%D9%83%D9%86%D8%AF%D9%8A_-_%D8%A7%D9%84%D9%81%D9%84%D8%B3%D9%81%D8%A9_%D8%A7%D9%84%D8%A3%D9%88%D9%84%D9%89) | [Creative Commons Attribution-ShareAlike License](https://creativecommons.org/licenses/by-sa/4.0/) |