@taj-p 2024-11-28
> Thanks! This looks great to me.

Thanks @grovesNL for the speedy attention 🙏 .

> Could you include a reference for the samples and ensure they're licensed in a way that we can include them here? Otherwise we could create our own sample.

Yes, definitely - thank you for the good suggestion and catch. I've updated the sources in https://github.com/grovesNL/glyphon/pull/121/commits/093eed717dd94c78c7a0943d7d36c39d41002e4f and attributed them in `samples/README.md`.